autoload added
git-svn-id: http://equinox:10080/_svn_root/AObjectServer@289 1a28b8b1-d4e8-a449-a3f8-e3d44feecac6

+++ AObjectServer/create_aos_distribution.py
@@ -143,7 +143,8 @@ syncPath(os.path.join(base_path, "AObjectServer", "CHANGELOG.txt"), TARGET_PATH)
 syncPath(os.path.join(base_path, "AObjectServer", "LICENSE.txt"), TARGET_PATH);
 syncPath(os.path.join(base_path, "AObjectServer", "CONTRIB.txt"), TARGET_PATH);
 syncPath(os.path.join(base_path, "AObjectServer", "AObjectServerWithoutALibrary.sln"), os.path.join(TARGET_PATH, "AObjectServer"));
-os.remove(os.path.join(TARGET_PATH, "AObjectServer", "AObjectServer.sln"));
+if (os.path.exists(os.path.join(TARGET_PATH, "AObjectServer", "AObjectServer.sln"))):
+  os.remove(os.path.join(TARGET_PATH, "AObjectServer", "AObjectServer.sln"));
 os.rename(os.path.join(TARGET_PATH, "AObjectServer", "AObjectServerWithoutALibrary.sln"), os.path.join(TARGET_PATH, "AObjectServer", "AObjectServer.sln"))
 
 print "----------AObjectServer--------";
